refactor branch code to separate module
